Fix the bug of empty service_instance_guid
This patch is to fix the bug introduced in PR 490.
We changed to store ServiceInstanceRequirement in the command instead
of the ServiceInstanceGuid because it is set as an empty string before
calling of ServiceInstanceRequirement.Execute(). Now after the command
runs, a correct guid can be get by using GetServiceInstance() of
ServiceInstanceRequirement.

we also found the delete service key command does not need Requirement
of ServiceInstance because it will also generate duplicate API call and
a warning message is desired instead of a failure message. So we remove it.

[#98066620]

Signed-off-by: Hua Zhang <zhuadl@cn.ibm.com>